+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- `modtype::preset::u64::thread_local::{with_modulus, F}`.
+
 ### Changed
 
 - `moving_ops_for_ref` to `no_impl_for_ref`.